refactor(workbench): extract compact overlay and appDialog host
Keep send/settings side effects in AppWorkbench via explicit callbacks instead of a catch-all ctx bag.
